Factor server code out of conftest to avoid circular dependency...
This was causing problems when testing Astropy when ASDF is installed.
