+++ asdf/conftest.py
@@ -3,6 +3,8 @@
 
 from __future__ import absolute_import, division, unicode_literals, print_function
 
+import os
+
 # this contains imports plugins that configure py.test for asdf tests.
 # by importing them here in conftest.py they are discoverable by py.test
 # no matter how it is invoked within the source tree.
@@ -10,20 +12,14 @@ from __future__ import absolute_import, division, unicode_literals, print_functi
 from astropy import __version__ as astropy_version
 from astropy.tests.pytest_plugins import *
 
-import multiprocessing
-import os
-import shutil
-import tempfile
-
 import pytest
 
 import six
 
-from .extern.RangeHTTPServer import RangeHTTPRequestHandler
-
 # This is to figure out the affiliated package version, rather than
 # using Astropy's
 from . import version
+from .tests.httpserver import HTTPServer, RangeHTTPServer
 
 packagename = os.path.basename(os.path.dirname(__file__))
 TESTED_VERSIONS[packagename] = version.version
@@ -46,55 +42,6 @@ try:
     del PYTEST_HEADER_MODULES['Scipy']
 except (NameError, KeyError):
     pass
-
-
-
-
-def run_server(queue, tmpdir, handler_class):  # pragma: no cover
-    """
-    Runs an HTTP server serving files from given tmpdir in a separate
-    process.  When it's ready, it sends a URL to the server over a
-    queue so the main process (the HTTP client) can start making
-    requests of it.
-    """
-    class HTTPRequestHandler(handler_class):
-        def translate_path(self, path):
-            path = handler_class.translate_path(self, path)
-            path = os.path.join(
-                tmpdir,
-                os.path.relpath(path, os.getcwd()))
-            return path
-
-    server = six.moves.socketserver.TCPServer(
-        ("127.0.0.1", 0), HTTPRequestHandler)
-    domain, port = server.server_address
-    url = "http://{0}:{1}/".format(domain, port)
-
-    queue.put(url)
-
-    server.serve_forever()
-
-
-class HTTPServer(object):
-    handler_class = six.moves.SimpleHTTPServer.SimpleHTTPRequestHandler
-
-    def __init__(self):
-        self.tmpdir = tempfile.mkdtemp()
-
-        q = multiprocessing.Queue()
-        self.process = multiprocessing.Process(
-            target=run_server,
-            args=(q, self.tmpdir, self.handler_class))
-        self.process.start()
-        self.url = q.get()
-
-    def finalize(self):
-        self.process.terminate()
-        shutil.rmtree(self.tmpdir)
-
-
-class RangeHTTPServer(HTTPServer):
-    handler_class = RangeHTTPRequestHandler
 
 
 @pytest.fixture()
